Turn @planetarium/cli npm pkg into ESM
