+++ Libplanet.Tools/.gitignore
@@ -1,5 +1,8 @@
 # npm
+.package.json.bak
+.package.json.tmp
 !bin/
 bin/*/
 node_modules/
 package-lock.json
+planetarium-cli-*.tgz
